This method does not work, a new branch is needed
The problem with this general approach is that worker threads are
calling the renderer when drawing a pixel; however, the renderer is not
thread safe. I tried multiple different ways to get around this, but
they all come back to the same fundamental issue.

Instead, every thread needs to modify a color array, and then the main
thread will merge that array into the display. This will likely be done
with surface or renderCopy.