Odd problems still occuring, could be closer
Added destructor for panel, reorganized the main function.
Renders properly when I include the if statement waiting for the
taskQueue to be empty (and subsequently the toberendered queue full),
this does not make any sense to me because the queues should be entirely
seperate. Also still running into thread dominance issues. Will be
trying this on valgrind and address sanitize on a linux os to see if
this catches anything.